+++ src/doomstat.h
@@ -670,6 +670,7 @@ extern int comp_cgunnersfx;
 extern int comp_flamst;
 extern int comp_keynoway;
 extern int comp_godface;
+extern int comp_powerrunout;
 extern int comp_deadoof;
 extern int comp_unusedpals;
 extern int comp_keypal;